pkg count with dpkg
Modified code to count pkgs with dpkg instead of apt. This is done in order to avoid the warning "WARNING: apt does not have a stable CLI interface. Use with caution in scripts.". 

"dpkg --get-selection" with no args returns all installed packages (but also those set to "deinstall", reason for the next grep)

Reference https://askubuntu.com/questions/445384/what-is-the-difference-between-apt-and-apt-get/446484#446484